+++ Platform/FO-DICOM.AspNetCore/Server/GeneralPurposeDicomService.cs
@@ -1,13 +1,12 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using FellowOakDicom.Imaging.Codec;
-using FellowOakDicom.Log;
 using FellowOakDicom.Network;
 using System;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
 
 namespace FellowOakDicom.AspNetCore.Server
 {
@@ -23,7 +22,7 @@ namespace FellowOakDicom.AspNetCore.Server
 
         private readonly DicomTransferSyntax[] _availableImageTransferSyntaxes;
 
-        public GeneralPurposeDicomService(INetworkStream stream, Encoding fallbackEncoding, Logger log, DicomServiceDependencies dependencies)
+        public GeneralPurposeDicomService(INetworkStream stream, Encoding fallbackEncoding, ILogger log, DicomServiceDependencies dependencies)
            : base(stream, fallbackEncoding, log, dependencies)
         {
             _availableImageTransferSyntaxes = DicomTransferSyntax.KnownEntries