+++ Tests/FO-DICOM.Tests/Network/AsyncDicomCStoreProviderTests.cs
@@ -5,11 +5,10 @@ using System;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FellowOakDicom.Imaging.Codec;
-using FellowOakDicom.Log;
 using FellowOakDicom.Network;
 using FellowOakDicom.Network.Client;
 using FellowOakDicom.Tests.Helpers;
+using Microsoft.Extensions.Logging;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -118,7 +117,7 @@ namespace FellowOakDicom.Tests.Network
                DicomTransferSyntax.JPEGProcess2_4,
         };
 
-        public AsyncDicomCStoreProviderPreferingUncompressedTS(INetworkStream stream, Encoding fallbackEncoding, Logger log, DicomServiceDependencies dependencies)
+        public AsyncDicomCStoreProviderPreferingUncompressedTS(INetworkStream stream, Encoding fallbackEncoding, ILogger log, DicomServiceDependencies dependencies)
             : base(stream, fallbackEncoding, log, dependencies)
         { }
 
@@ -144,7 +143,7 @@ namespace FellowOakDicom.Tests.Network
 
     public class AsyncDicomCStoreProvider : DicomService, IDicomServiceProvider, IDicomCStoreProvider
     {
-        public AsyncDicomCStoreProvider(INetworkStream stream, Encoding fallbackEncoding, Logger log,
+        public AsyncDicomCStoreProvider(INetworkStream stream, Encoding fallbackEncoding, ILogger log,
             DicomServiceDependencies dependencies)
             : base(stream, fallbackEncoding, log, dependencies)
         {