+++ Tests/FO-DICOM.Tests/Network/Client/DicomClientTest.cs
@@ -1,13 +1,14 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using FellowOakDicom.Log;
 using FellowOakDicom.Network;
 using FellowOakDicom.Network.Client;
 using FellowOakDicom.Network.Client.Advanced.Connection;
 using FellowOakDicom.Network.Client.States;
 using FellowOakDicom.Tests.Helpers;
+using FellowOakDicom.Tests.Log;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
@@ -147,14 +148,14 @@ namespace FellowOakDicom.Tests.Network.Client
         public async Task LogAssociationProperties()
         {
             var writer = new StringWriter();
-            ILogManager logManager = new TextWriterLogManager(writer);
+            var logger = new TextWriterLogger(writer);
 
             int port = Ports.GetNext();
             using (CreateServer<DicomCEchoProvider>(port))
             {
                 var request = new DicomCEchoRequest { };
                 var client = CreateClient("127.0.0.1", port, false, "LOG-SCU", "ANY-SCP");
-                client.Logger = logManager.GetLogger("client");
+                client.Logger = logger;
 
                 await client.AddRequestAsync(request).ConfigureAwait(false);
                 await client.SendAsync().ConfigureAwait(false);
@@ -180,7 +181,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 // but in case some developer registeres a custom Factory or creates DicomClient directly for some other reason.
                 var client = new DicomClient("localhost", port, false, "STORAGECOMMITTEST", "DE__257a276f6d47",
                     new DicomClientOptions { }, new DicomServiceOptions { },
-                    Setup.ServiceProvider.GetRequiredService<ILogManager>().GetLogger("DicomClient"),
+                    Setup.ServiceProvider.GetRequiredService<ILoggerFactory>(),
                     Setup.ServiceProvider.GetRequiredService<IAdvancedDicomClientConnectionFactory>());
                 await client.AddRequestAsync(request).ConfigureAwait(false);
 
@@ -445,7 +446,7 @@ namespace FellowOakDicom.Tests.Network.Client
             catch (Exception ex)
             {
                 result = false;
-                client.Logger.Error("Send failed, exception: {0}", ex);
+                client.Logger.LogError("Send failed, exception: {0}", ex);
                 awaiter.Set();
             }
 
@@ -1013,7 +1014,7 @@ namespace FellowOakDicom.Tests.Network.Client
             var logger = _logger.IncludePrefix("Responses");
             foreach (var r in responses)
             {
-                logger.Info($"{r.Type} [{r.RequestMessageID}]: " +
+                logger.LogInformation($"{r.Type} [{r.RequestMessageID}]: " +
                             $"Status = {r.Status.State}, " +
                             $"Code = {r.Status.Code}, " +
                             $"ErrorComment = {r.Status.ErrorComment}, " +
@@ -1036,7 +1037,7 @@ namespace FellowOakDicom.Tests.Network.Client
             var client = CreateClient("127.0.0.1", port, false, "SCU", "ANY-SCP");
             client.ClientOptions.AssociationLingerTimeoutInMs = lingerTimeoutInSeconds * 1000;
 
-            logger.Info($"Beginning {numberOfRequests} parallel requests with {secondsBetweenEachRequest}s between each request");
+            logger.LogInformation($"Beginning {numberOfRequests} parallel requests with {secondsBetweenEachRequest}s between each request");
 
             var responses = new ConcurrentBag<DicomCEchoResponse>();
             var sendTasks = new List<Task>();
@@ -1055,7 +1056,7 @@ namespace FellowOakDicom.Tests.Network.Client
                     sendTasks.Add(client.SendAsync());
                 }
 
-                logger.Info($"Waiting {secondsBetweenEachRequest} seconds between requests");
+                logger.LogInformation($"Waiting {secondsBetweenEachRequest} seconds between requests");
                 await Task.Delay(TimeSpan.FromSeconds(secondsBetweenEachRequest)).ConfigureAwait(false);
             }
 
@@ -1086,7 +1087,7 @@ namespace FellowOakDicom.Tests.Network.Client
             var client = CreateClient("127.0.0.1", port, false, "SCU", "ANY-SCP");
             client.ClientOptions.AssociationLingerTimeoutInMs = lingerTimeoutInSeconds * 1000;
 
-            logger.Info($"Beginning {numberOfRequests} parallel requests with {secondsBetweenEachRequest}s between each request");
+            logger.LogInformation($"Beginning {numberOfRequests} parallel requests with {secondsBetweenEachRequest}s between each request");
 
             var responses = new ConcurrentBag<DicomCEchoResponse>();
 
@@ -1106,7 +1107,7 @@ namespace FellowOakDicom.Tests.Network.Client
                     sendTasks.Add(client.SendAsync());
                 }
 
-                logger.Info($"Waiting {secondsBetweenEachRequest} seconds between requests");
+                logger.LogInformation($"Waiting {secondsBetweenEachRequest} seconds between requests");
                 await Task.Delay(TimeSpan.FromSeconds(secondsBetweenEachRequest)).ConfigureAwait(false);
             }
             await Task.WhenAll(sendTasks).ConfigureAwait(false);
@@ -1138,7 +1139,7 @@ namespace FellowOakDicom.Tests.Network.Client
             var client = CreateClient("127.0.0.1", port, false, "SCU", "ANY-SCP");
             client.ClientOptions.AssociationLingerTimeoutInMs = lingerTimeoutInSeconds * 1000;
 
-            logger.Info($"Beginning {numberOfRequests} parallel requests with {secondsBetweenEachRequest}s between each request");
+            logger.LogInformation($"Beginning {numberOfRequests} parallel requests with {secondsBetweenEachRequest}s between each request");
 
             var responses = new ConcurrentBag<DicomCEchoResponse>();
 
@@ -1158,7 +1159,7 @@ namespace FellowOakDicom.Tests.Network.Client
                     sendTasks.Add(client.SendAsync());
                 }
 
-                logger.Info($"Waiting {secondsBetweenEachRequest} seconds between requests");
+                logger.LogInformation($"Waiting {secondsBetweenEachRequest} seconds between requests");
                 await Task.Delay(TimeSpan.FromSeconds(secondsBetweenEachRequest)).ConfigureAwait(false);
             }
             await Task.WhenAll(sendTasks).ConfigureAwait(false);
@@ -1193,7 +1194,7 @@ namespace FellowOakDicom.Tests.Network.Client
             var client = CreateClient("127.0.0.1", port, false, "SCU", "ANY-SCP");
             client.ClientOptions.AssociationLingerTimeoutInMs = lingerTimeoutInSeconds * 1000;
 
-            logger.Info($"Beginning {numberOfRequests} parallel requests with variable wait times between each request");
+            logger.LogInformation($"Beginning {numberOfRequests} parallel requests with variable wait times between each request");
 
             var responses = new ConcurrentBag<DicomCEchoResponse>();
 
@@ -1205,7 +1206,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 {
                     OnResponseReceived = (req, res) => responses.Add(res)
                 };
-                logger.Info($"Adding request {i}");
+                logger.LogInformation($"Adding request {i}");
                 await client.AddRequestAsync(request).ConfigureAwait(false);
 
                 if (client.IsSendRequired)
@@ -1215,7 +1216,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 }
 
                 var secondsToWait = secondsBetweenEachRequest[i];
-                logger.Info($"Waiting {secondsToWait} seconds between requests");
+                logger.LogInformation($"Waiting {secondsToWait} seconds between requests");
                 await Task.Delay(TimeSpan.FromSeconds(secondsToWait)).ConfigureAwait(false);
             }
             await Task.WhenAll(sendTasks).ConfigureAwait(false);
@@ -1247,7 +1248,7 @@ namespace FellowOakDicom.Tests.Network.Client
 
                 client.OnCStoreRequest = async request =>
                 {
-                    logger.Info("Handling C-Store request");
+                    logger.LogInformation("Handling C-Store request");
                     capturedCStoreRequest = request;
                     await Task.Delay(50).ConfigureAwait(false);
                     return new DicomCStoreResponse(request, DicomStatus.Success);
@@ -1257,7 +1258,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 var seriesInstanceUID = "999.999.94827453";
                 var sopInstanceUID = "999.999.133.1996.1.1800.1.6.21";
 
-                logger.Info("Sending C-Get request");
+                logger.LogInformation("Sending C-Get request");
                 await client.AddRequestAsync(new DicomCGetRequest(studyInstanceUID, seriesInstanceUID, sopInstanceUID)).ConfigureAwait(false);
 
                 var pcs = DicomPresentationContext.GetScpRolePresentationContextsFromStorageUids(
@@ -1290,7 +1291,7 @@ namespace FellowOakDicom.Tests.Network.Client
             client.ClientOptions.MaximumNumberOfRequestsPerAssociation = maxRequestsPerAssoc;
             client.NegotiateAsyncOps(10, 10);
 
-            logger.Info($"Beginning {numberOfRequests} requests with max {maxRequestsPerAssoc} requests / association");
+            logger.LogInformation($"Beginning {numberOfRequests} requests with max {maxRequestsPerAssoc} requests / association");
 
             var responses = new ConcurrentBag<DicomCEchoResponse>();
             for (var i = 1; i <= numberOfRequests; i++)
@@ -1300,7 +1301,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 {
                     OnResponseReceived = (request, response) =>
                     {
-                        logger.Debug($"Request completed: {iLocal}");
+                        logger.LogDebug($"Request completed: {iLocal}");
                         responses.Add(response);
                     }
                 };
@@ -1339,7 +1340,7 @@ namespace FellowOakDicom.Tests.Network.Client
             }
             catch (Exception exception)
             {
-                _logger.Error("Error occurred during send {e}", exception);
+                _logger.LogError("Error occurred during send {e}", exception);
                 capturedException = exception;
             }
             Assert.NotNull(capturedException);
@@ -1365,18 +1366,18 @@ namespace FellowOakDicom.Tests.Network.Client
                 {
                     OnResponseReceived = (request, response) =>
                     {
-                        logger.Info("Received echo response 1, disposing server");
+                        logger.LogInformation("Received echo response 1, disposing server");
                         echoResponse1 = response;
                         // ReSharper disable once AccessToDisposedClosure This is an edge case we are trying to test
                         server?.Dispose();
-                        logger.Info("Server is disposed");
+                        logger.LogInformation("Server is disposed");
                     }
                 };
                 var echoRequest2 = new DicomCEchoRequest
                 {
                     OnResponseReceived = (request, response) =>
                     {
-                        logger.Info("Received echo response 2");
+                        logger.LogInformation("Received echo response 2");
                         echoResponse2 = response;
                     }
                 };
@@ -1384,7 +1385,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 {
                     OnResponseReceived = (request, response) =>
                     {
-                        logger.Info("Received echo response 3");
+                        logger.LogInformation("Received echo response 3");
                         echoResponse3 = response;
                     }
                 };
@@ -1419,7 +1420,7 @@ namespace FellowOakDicom.Tests.Network.Client
 
         public class MockCEchoProvider : DicomService, IDicomServiceProvider, IDicomCEchoProvider
         {
-            public MockCEchoProvider(INetworkStream stream, Encoding fallbackEncoding, Logger log,
+            public MockCEchoProvider(INetworkStream stream, Encoding fallbackEncoding, ILogger log,
                 DicomServiceDependencies dependencies)
                 : base(stream, fallbackEncoding, log, dependencies)
             {
@@ -1475,7 +1476,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 DicomTransferSyntax.ExplicitVRLittleEndian
             };
 
-            public ExplicitLECStoreProvider(INetworkStream stream, Encoding fallbackEncoding, Logger log,
+            public ExplicitLECStoreProvider(INetworkStream stream, Encoding fallbackEncoding, ILogger log,
                 DicomServiceDependencies dependencies)
                 : base(stream, fallbackEncoding, log, dependencies)
             {
@@ -1675,7 +1676,7 @@ namespace FellowOakDicom.Tests.Network.Client
                 }
                 catch (Exception e)
                 {
-                    Logger.Error("Could not send file via C-Store request: {error}", e);
+                    Logger.LogError("Could not send file via C-Store request: {error}", e);
                     result = new DicomCGetResponse(request, DicomStatus.ProcessingFailure);
                 }