+++ Tests/FO-DICOM.Tests/Network/DicomNEventReportResponseTest.cs
@@ -7,8 +7,7 @@ using System;
 using System.Text;
 using System.Threading.Tasks;
 using FellowOakDicom.Network.Client;
-using FellowOakDicom.Log;
-using FellowOakDicom.Imaging.Codec;
+using Microsoft.Extensions.Logging;
 
 namespace FellowOakDicom.Tests.Network
 {