+++ Tests/FO-DICOM.Tests/Network/SimpleCStoreProvider.cs
@@ -1,14 +1,13 @@
 ﻿// Copyright (c) 2012-2021 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using FellowOakDicom.Log;
 using FellowOakDicom.Network;
 using System;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
-using FellowOakDicom.Imaging.Codec;
 using System.Collections.Generic;
+using Microsoft.Extensions.Logging;
 
 namespace FellowOakDicom.Tests.Network
 {
@@ -45,7 +44,7 @@ namespace FellowOakDicom.Tests.Network
             DicomTransferSyntax.ImplicitVRLittleEndian
         };
 
-        public SimpleCStoreProvider(INetworkStream stream, Encoding fallbackEncoding, Logger log,
+        public SimpleCStoreProvider(INetworkStream stream, Encoding fallbackEncoding, ILogger log,
             DicomServiceDependencies dependencies)
             : base(stream, fallbackEncoding, log, dependencies)
         {
@@ -78,7 +77,7 @@ namespace FellowOakDicom.Tests.Network
         public async Task<DicomCStoreResponse> OnCStoreRequestAsync(DicomCStoreRequest request)
         {
             var tempName = Path.GetTempFileName();
-            Logger.Info(tempName);
+            Logger.LogInformation(tempName);
 
             // as the incoming file could be not-valid
             request.File.Dataset.ValidateItems = false;